Rm poetry build script, save setup at build for reference and local builds
