+++ pyproject.toml
@@ -28,9 +28,6 @@ flake8 = "^3.9.2"
 numpy = "^1.21.0"
 pandas = "^1.3.0"
 
-[tool.poetry.build]
-script = "build.py"
-
 [build-system]
 requires = [
   "poetry-core>=1.0.0",
